requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.1.7
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
entrypoints==0.3
future==0.18.2
gunicorn==19.9.0
hiredis==2.0.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
ipaddr==2.2.0
itypes==1.2.0
jeepney==0.6.0
Jinja2==2.10.3
keyring==18.0.1
linecache2==1.0.0
lockfile==0.12.2
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==2.0.3
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.3.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyotp==2.4.1
pyparsing==2.4.6
PyQRCode==1.2.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
SecretStorage==3.3.1
selenium==3.141.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
traceback2==1.4.0
twilio==6.50.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
unittest2==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
zope.interface==5.4.0
